(BSR)[PRO] refactor: remove adapters
