--- pro/src/pages/CollectiveBookings/adapters/getUserHasCollectiveBookingsAdapter.ts
@@ -1,7 +1,0 @@
-import { api } from 'apiClient/api'
-
-export const getUserHasCollectiveBookingsAdapter = async () => {
-  const { hasBookings } = await api.getUserHasCollectiveBookings()
-
-  return hasBookings
-}
